requirements.txt
pillow == 10.2.0
torch == 2.2.2
numpy == 1.26.4
scikit-learn == 1.2.2
torchvision == 0.17.2
matplotlib == 3.9.1
segmentation_models_pytorch == 0.3.3
tqdm == 4.65.0
opencv-python == 4.10.0.84